Refactor math types: Introduce Bool2 and Bool3 structs, replace Point2D/Point3D with Vec2/Vec3, and remove point.rs
- Added Bool2 and Bool3 structs for component-wise boolean operations.
- Updated Bounds2D and Bounds3D to use Vec2 and Vec3 instead of custom Point types.
- Removed point.rs as its functionality is now covered by Vec2 and Vec3.
- Simplified vector operations by leveraging existing Bevy methods.
- Enhanced the vector module with additional utility functions and traits.

+++ src/math/probability/seed/events.rs
@@ -2,5 +2,5 @@ use bevy::prelude::*;
 
 #[derive(Event, Debug, Clone)]
 pub struct SeedChangedEvent {
-    pub new_seed: u64,
+    pub new_seed: u32,
 }
